refactor(identifiers): split Layer 3 builder/government.rs into per-country submodules
Slice of the #193 code-health umbrella — file-length category, Layer 3.

identifiers/builder/government.rs had grown to 1902 LOC. Split into a
government/ directory of 20 files mirroring PR #317's primitive-layer
split, so Layer 1 (primitives/identifiers/government/detection/*) and
Layer 3 (identifiers/builder/government/*) now have one-for-one
per-country file parity.

  government/mod.rs          (struct + metric_names + ctors + 4 tests, 179 LOC)
  government/india.rs        (Aadhaar, PAN, GSTIN, vehicle-reg, voter-id, passport, 319 LOC)
  government/europe.rs       (Spain NIF/NIE, Italy CF, Finland HETU, Poland PESEL, 220 LOC)
  government/brazil.rs       (CPF + CNPJ, 122 LOC)
  government/ssn.rs          (SSN, 115 LOC)
  government/australia.rs    (TFN + ABN, 89 LOC)
  government/uk_ni.rs        (UK NINO, 86 LOC)
  government/aggregate.rs    (cross-cutting find_all / is_government_present, 77 LOC)
  government/tax_id.rs       (Tax ID + EIN, 78 LOC)
  government/national_id.rs  (generic + Canada SIN, 75 LOC)
  government/vehicle_id.rs   (VIN, 74 LOC)
  + 9 single-identifier files (driver_license, passport, mexico, nigeria,
    thailand, singapore, korea, test_patterns, cache; 44-65 LOC each)

Approach: Rust permits multiple `impl GovernmentBuilder` blocks for the
same type across files. The struct, `metric_names` private module
(promoted to `pub(super)` so siblings can see it), and the three
constructors stay in `mod.rs`; each per-country/domain submodule adds one
`impl GovernmentBuilder { ... }` block holding that section's methods.

Pure file re-organisation:
  - Zero public-API changes — identifiers/shortcuts/government.rs imports
    `GovernmentBuilder` by name through the existing pub-use re-export,
    which still resolves correctly through a directory module
  - No method-signature, behaviour, or observe-instrumentation changes;
    every method body — including `Instant::now()`, `record(...)`,
    `increment_by(...)`, `observe::debug/warn` calls — is moved
    byte-for-byte
  - The 4 builder-construction tests stay in `mod.rs`
  - Largest new file is 319 LOC (india.rs), well under the 500-LOC
    warning threshold

Verified via just preflight: clippy + arch-check + 6650-test octarine
suite all green.

Refs #193
