+++ crates/octarine/src/identifiers/builder/government/tax_id.rs
@@ -1,0 +1,78 @@
+//! Tax ID and EIN methods.
+
+use super::*;
+
+impl GovernmentBuilder {
+    /// Check if value is a tax ID
+    #[must_use]
+    pub fn is_tax_id(&self, value: &str) -> bool {
+        self.inner.is_tax_id(value)
+    }
+
+    /// Find all tax IDs in text
+    #[must_use]
+    pub fn find_tax_ids_in_text(&self, text: &str) -> Vec<IdentifierMatch> {
+        self.inner.find_tax_ids_in_text(text)
+    }
+
+    /// Check if value is a valid EIN (Employer Identification Number)
+    ///
+    /// Validates both the `XX-XXXXXXX` format and the IRS campus code prefix.
+    #[must_use]
+    pub fn is_ein(&self, value: &str) -> bool {
+        self.inner.is_ein(value)
+    }
+
+    /// Find all valid EINs in text
+    #[must_use]
+    pub fn find_eins_in_text(&self, text: &str) -> Vec<IdentifierMatch> {
+        self.inner.find_eins_in_text(text)
+    }
+
+    /// Validate EIN format
+    ///
+    /// # Errors
+    ///
+    /// Returns `Problem` if the EIN format is invalid
+    pub fn validate_ein(&self, ein: &str) -> Result<(), Problem> {
+        self.inner.validate_ein(ein)
+    }
+
+    /// Redact a tax ID with explicit strategy
+    #[must_use]
+    pub fn redact_tax_id_with_strategy(
+        &self,
+        tax_id: &str,
+        strategy: TaxIdRedactionStrategy,
+    ) -> String {
+        self.inner.redact_tax_id_with_strategy(tax_id, strategy)
+    }
+
+    /// Redact all tax IDs in text with explicit strategy
+    #[must_use]
+    pub fn redact_tax_ids_in_text_with_strategy(
+        &self,
+        text: &str,
+        strategy: TaxIdRedactionStrategy,
+    ) -> String {
+        self.inner
+            .redact_tax_ids_in_text_with_strategy(text, strategy)
+    }
+
+    /// Normalize an EIN (remove formatting)
+    #[must_use]
+    pub fn normalize_ein(&self, ein: &str) -> String {
+        self.inner.normalize_ein(ein)
+    }
+
+    /// Convert EIN to standard hyphenated format
+    #[must_use]
+    pub fn to_ein_with_hyphen(&self, ein: &str) -> String {
+        self.inner.to_ein_with_hyphen(ein)
+    }
+
+    /// Sanitize an EIN (normalize + validate)
+    pub fn sanitize_ein(&self, ein: &str) -> Result<String, Problem> {
+        self.inner.sanitize_ein(ein)
+    }
+}
